* MapGenerator > __generate_separated_planets(): made finite loop based on condition instead of infinite loop with "break"s; * other minor fixes.

+++ utils/utils.py
@@ -64,7 +64,7 @@ class EventPriorityQueue(object):
 class Coords(object):
     """ координата точки """
 
-    def __init__(self, x: int = 0, y: int = 0):
+    def __init__(self, x: float = 0, y: float = 0):
         self.x = x
         self.y = y
